+++ src/extension-test/integration/suite/annotations-test.ts
@@ -24,8 +24,6 @@ suite('Annotations suite', () => {
   });
 
   test('decorate result trims leading whitespaces', async function () {
-    testUtil.log(suite, 'activeEditor');
-
     // any file would do
     const testFilePath = path.join(testUtil.testDataDir, 'test.clj');
     const editor = await testUtil.openFile(testFilePath);
@@ -62,5 +60,77 @@ suite('Annotations suite', () => {
 
     await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
     testUtil.log(suite, 'test.clj closed');
+  });
+
+  test('tooltip contents of evaluated result', async function () {
+    // any file would do
+    const testFilePath = path.join(testUtil.testDataDir, 'test.clj');
+    const editor = await testUtil.openFile(testFilePath);
+
+    // Any range will do, the range below happens to be around "hover-map".
+    const selection = new vscode.Selection(2, 5, 2, 14);
+    editor.selection = selection;
+    const proxyEditor = testUtil.createVscTextEditorProxy(editor);
+
+    // will eventually call on the editor's `setDecorations` fn with the text to render.
+    const setDecorationsSpy = sinon.spy(proxyEditor, 'setDecorations');
+
+    const resultString = ' :Line1 with a leading space\n:Line2 some text\n\nLine4 the end.';
+
+    // SUCCESS case, format should contain the result in a clojure code block
+    {
+      annotations.decorateSelection(
+        resultString,
+        selection,
+        proxyEditor,
+        null,
+        annotations.AnnotationStatus.SUCCESS
+      );
+
+      //await testUtil.sleep(60000);
+
+      assert.deepStrictEqual(
+        setDecorationsSpy.firstCall.args[0],
+        annotations._getEvalSelectionDecorationTypes(annotations.AnnotationStatus.SUCCESS)
+      );
+      assert.deepStrictEqual(setDecorationsSpy.firstCall.args[1], []);
+      assert.deepStrictEqual(
+        setDecorationsSpy.secondCall.args[0],
+        annotations._getEvalSelectionDecorationTypes(annotations.AnnotationStatus.SUCCESS)
+      );
+      const decorateOpts = setDecorationsSpy.secondCall.args[1] as vscode.DecorationOptions[];
+      const hoverMessage = decorateOpts[0].hoverMessage as vscode.MarkdownString;
+      const expectedResultSuccess = `${annotations._getDecorateSelectionHeader(
+        resultString
+      )}\n\`\`\`clojure\n${resultString}\n\`\`\``;
+      assert.strictEqual(hoverMessage.value, expectedResultSuccess);
+    }
+
+    sinon.reset();
+
+    // ERROR case, should contain the result in a plain code block (```)
+    {
+      annotations.decorateSelection(
+        resultString,
+        selection,
+        proxyEditor,
+        null,
+        annotations.AnnotationStatus.ERROR
+      );
+
+      assert.deepStrictEqual(
+        setDecorationsSpy.firstCall.args[0],
+        annotations._getEvalSelectionDecorationTypes(annotations.AnnotationStatus.ERROR)
+      );
+      assert.deepStrictEqual(setDecorationsSpy.firstCall.args[1], []);
+      const decorateOpts = setDecorationsSpy.secondCall.args[1] as vscode.DecorationOptions[];
+      const hoverMessage = decorateOpts[0].hoverMessage as vscode.MarkdownString;
+      const expectedResultSuccess = `${annotations._getDecorateSelectionHeader(
+        resultString
+      )}\n\`\`\`\n${resultString}\n\`\`\``;
+      assert.strictEqual(hoverMessage.value, expectedResultSuccess);
+    }
+
+    await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
   });
 });